Struct zng_wgt_progress::Progress

pub struct Progress { /* private fields */ }

Status update about a task progress.

Implementations

impl Progress

pub fn indeterminate() -> Progress

New indeterminate.

pub fn complete() -> Progress

New completed.

pub fn from_fct(factor: impl Into<Factor>) -> Progress

New with a factor of completion.

The factor must be in the 0..=1 range, with a rounding error of 0.001, values outside this range are converted to indeterminate.

pub fn from_n_of(n: usize, total: usize) -> Progress

New with completed n of total.

pub fn with_msg(self, msg: impl Into<Txt>) -> Progress

Set the display message about the task status update.

pub fn with_meta_mut( self, meta: impl FnOnce(StateMapMut<'_, Progress>), ) -> Progress

Set custom status metadata for writing.

Note that metadata is shared between all clones of self.

pub fn and_fct(self, factor: impl Into<Factor>) -> Progress

Combine the factor completed fct with another factor.

pub fn and_n_of(self, n: usize, total: usize) -> Progress

Combine the factor completed fct with another factor computed from n of total.

pub fn with_fct(self, factor: impl Into<Factor>) -> Progress

Replace the fct value with a new factor.

pub fn with_n_of(self, n: usize, total: usize) -> Progress

Replace the fct value with a new factor computed from n of total.

pub fn fct(&self) -> Factor

Factor completed.

Is -1.fct() for indeterminate, otherwise is a value in the 0..=1 range, 1.fct() indicates task completion.

pub fn is_indeterminate(&self) -> bool

Factor of completion cannot be known.

pub fn is_complete(&self) -> bool

Task has completed.

pub fn msg(&self) -> Txt

Display text about the task status update.

pub fn with_meta<T>( &self, visitor: impl FnOnce(StateMapRef<'_, Progress>) -> T, ) -> T

Borrow the custom status metadata for reading.