Crate zng_wgt_progress

Progress indicator widget.

Crate

This crate is part of the zng project.

Structs

DefaultStyle

W Progress view default style (progress bar with message text).

Progress

Status update about a task progress.

ProgressView

W Progress indicator widget.

SimpleBarStyle

W Progress view style that is only the progress bar, no message text.

Statics

PROGRESS_VAR

The progress status value in a ProgressView

STYLE_FN_VAR

Contextual style variable.

Functions

collapse_complete

P Collapse visibility when Progress::is_complete.

is_indeterminate

P Getter property that is true when progress is indeterminate.

on_complete

P Event raised when progress updates to a complete state or inits completed.

on_progress

P Event raised for each progress update, and once after info init.

progress

P The progress status to be displayed.

style_fn

P Extends or replaces the widget style.